Replace incorrect keywords.txt
The keywords.txt in the root of the repository is for another library. The keywords.txt in the src folder is for this library. The Arduino IDE uses the keywords.txt in the root library folder so the correct keywords.txt file was moved to the root to replace the incorrect file.